GNAT-2440 Create and subscribe to topics programmatically

+++ src/main/java/com/equinix/edn/networkdatatransformer/config/kafka/KafkaTopicConfig.java
@@ -1,5 +1,8 @@
 package com.equinix.edn.networkdatatransformer.config.kafka;
 
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,5 +13,7 @@ import lombok.NoArgsConstructor;
 public class KafkaTopicConfig {
     private String topicName;
     private Integer partitions;
-    private Integer replicas;
+    private Short replicas;
+
+
 }
